backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-multipart==0.0.7
pydantic==2.6.4
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
httpx==0.27.0
pytesseract==0.3.10
Pillow==10.3.0
celery==5.3.6
redis==5.0.1
ftfy==6.1.1
